+++ packages/frontend/src/pages/theme-editor.vue
@@ -78,7 +78,7 @@ import tinycolor from 'tinycolor2';
 import JSON5 from 'json5';
 import lightTheme from '@@/themes/_light.json5';
 import darkTheme from '@@/themes/_dark.json5';
-import { host } from '@@/js/config.js';
+import { localHost } from '@@/js/config.js';
 import type { Theme } from '@/theme.js';
 import { genId } from '@/utility/id.js';
 import MkButton from '@/components/MkButton.vue';
@@ -128,7 +128,7 @@ const fgColors = [
 const theme = ref<Theme>({
 	id: genId(),
 	name: 'untitled',
-	author: `@${$i.username}@${toUnicode(host)}`,
+	author: `@${$i.username}@${toUnicode(localHost)}`,
 	base: 'light',
 	props: lightTheme.props,
 });